feat: Update PHP upload max filesize configuration
The code changes include adding the `PHP_UPLOAD_MAX_FILESIZE` environment variable to the `get-configuration` and `configure-module` actions. The default value is set to "100M". This change allows users to configure the maximum size of files that can be uploaded to the server.

Note: This commit message follows the established convention of using a prefix to indicate the type of change (feat for feature).